Add support to handle very large files.   (#21)
Closes #18

This PR adds a change in the way we handle target downloads. Previously we loaded the target into memory and then performed the requisite validations. Now we write the download to a temp file, using streams to validate hashes. If all validations pass we mv the temp file to the staging file.